implement cylinder rendering using polyline renderer protocol
